Fixes with usable items.
Fixed checks volume/weight items on transfers.

Fixed random critter move.



git-svn-id: https://cvet.by:8443/svn/FOnlineSource@86 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: cf8a6b08a6b0d84080cf78a30e13b7576269ef2d

+++ Source/CritterCl.cpp
@@ -387,7 +387,7 @@ DWORD CritterCl::GetItemsWeightKg()
 
 DWORD CritterCl::GetItemsVolume()
 {
-	WORD res=0;
+	DWORD res=0;
 	for(ItemPtrMapIt it=InvItems.begin();it!=InvItems.end();++it)
 		res+=(*it).second->GetVolume();
 	return res;
